+++ src/components/AddTodoForm.jsx
@@ -1,0 +1,35 @@
+
+import PropTypes from "prop-types";
+import { useState } from "react";
+import "../App.css";
+
+function AddTodoForm({ onAddTodo }) {
+  const [newTodo, setNewTodo] = useState("");
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (newTodo.trim() === "") return;
+    onAddTodo(newTodo);
+    setNewTodo("");
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="todo-form">
+      <input
+        type="text"
+        value={newTodo}
+        onChange={(e) => setNewTodo(e.target.value)}
+        placeholder="Enter new task..."
+        required
+      />
+      <button type="submit">Add</button>
+    </form>
+  );
+}
+
+AddTodoForm.propTypes = {
+  onAddTodo: PropTypes.func
+};
+
+
+export default AddTodoForm;
