+++ src/components/TodoList.jsx
@@ -1,3 +1,4 @@
+import PropTypes from "prop-types";
 import TodoListItem from "./TodoListItem";
 
 
@@ -16,4 +17,15 @@ function TodoList({todoList, onRemoveTodo}) {
       </ul> 
   )
 }
+
+TodoList.propTypes = {
+  todoList: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+      title: PropTypes.string.isRequired,
+    })
+  ).isRequired,
+  onRemoveTodo: PropTypes.func.isRequired,
+};
+
 export default TodoList;
